+++ Libs/Models/Gateway/GatewayUserAccount.cs
@@ -4,7 +4,7 @@
     {
         public int id { get; set; }
         public string name { get; set; }
-        public decimal balance { get; set; }
+        public double balance { get; set; }
         public bool canTrade { get; set; }
         public bool isVisible { get; set; }
         public bool simulated { get; set; }
