+++ Libs/Models/History/AggregateBarModel.cs
@@ -3,10 +3,10 @@
     public class AggregateBarModel
     {
         public DateTime t { get; set; }
-        public decimal o { get; set; }
-        public decimal h { get; set; }
-        public decimal l { get; set; }
-        public decimal c { get; set; }
+        public double o { get; set; }
+        public double h { get; set; }
+        public double l { get; set; }
+        public double c { get; set; }
         public long v { get; set; }
     }
 }
